refactor(ui): extract app window lifecycle helpers
